+++ test/tgenprop.c
@@ -47,7 +47,7 @@ int         prop1_def=10;   /* Property 1 default value */
 #define PROP1_DEF_VALUE (&prop1_def)
 
 #define PROP2_NAME     "Property 2"
-float         prop2_def=(float)3.14;   /* Property 2 default value */
+float         prop2_def=3.14F;   /* Property 2 default value */
 #define PROP2_SIZE      sizeof(prop2_def)
 #define PROP2_DEF_VALUE (&prop2_def)
 
@@ -57,7 +57,7 @@ char          prop3_def[10]="Ten chars";   /* Property 3 default value */
 #define PROP3_DEF_VALUE (&prop3_def)
 
 #define PROP4_NAME     "Property 4"
-double          prop4_def=1.41;   /* Property 4 default value */
+double          prop4_def=1.41F;   /* Property 4 default value */
 #define PROP4_SIZE      sizeof(prop4_def)
 #define PROP4_DEF_VALUE (&prop4_def)
 
@@ -683,7 +683,7 @@ test_genprop_basic_list(void)
     /* Verify the floating-poing value in this way to avoid compiler warning. */
     if(!FLT_ABS_EQUAL(prop2_value,*PROP2_DEF_VALUE))
 	printf("*** UNEXPECTED VALUE from %s should be %f, but is %f at line %4d in %s\n",
-	    "H5Pget", *PROP2_DEF_VALUE, prop2_value, (int)__LINE__, __FILE__);
+	    "H5Pget", (double)*PROP2_DEF_VALUE, (double)prop2_value, (int)__LINE__, __FILE__);
 
 
     /* Close list */
@@ -775,7 +775,7 @@ test_genprop_basic_list_prop(void)
     /* Verify the floating-poing value in this way to avoid compiler warning. */
     if(!FLT_ABS_EQUAL(prop2_value,*PROP2_DEF_VALUE))
 	printf("*** UNEXPECTED VALUE from %s should be %f, but is %f at line %4d in %s\n",
-	    "H5Pget", *PROP2_DEF_VALUE, prop2_value, (int)__LINE__, __FILE__);
+	    "H5Pget", (double)*PROP2_DEF_VALUE, (double)prop2_value, (int)__LINE__, __FILE__);
 
 
     /* Check values of temporary properties (set with regular values) */
@@ -1239,7 +1239,7 @@ test_genprop_list_callback(void)
     /* Verify the floating-poing value in this way to avoid compiler warning. */
     if(!FLT_ABS_EQUAL(prop2_value,*PROP2_DEF_VALUE))
 	printf("*** UNEXPECTED VALUE from %s should be %f, but is %f at line %4d in %s\n",
-	    "H5Pget", *PROP2_DEF_VALUE, prop2_value, (int)__LINE__, __FILE__);
+	    "H5Pget", (double)*PROP2_DEF_VALUE, (double)prop2_value, (int)__LINE__, __FILE__);
 
     /* Check values of temporary properties (set with regular values) */
     ret = H5Pget(lid1, PROP3_NAME,&prop3_value);
